Extract Nitrokey Storage only features to separate unit
