+++ NK_C_API_storage.h
@@ -1,0 +1,6 @@
+#ifndef LIBNITROKEY_NK_C_API_STORAGE_H
+#define LIBNITROKEY_NK_C_API_STORAGE_H
+
+
+
+#endif // LIBNITROKEY_NK_C_API_STORAGE_H
