+++ NitrokeyManagerStorage.h
@@ -1,0 +1,6 @@
+#ifndef LIBNITROKEY_NITROKEYMANAGERSTORAGE_H
+#define LIBNITROKEY_NITROKEYMANAGERSTORAGE_H
+
+#include "NitrokeyManager.h"
+
+#endif // LIBNITROKEY_NITROKEYMANAGERSTORAGE_H
