+++ libnitrokey/nk_strndup.h
@@ -1,0 +1,9 @@
+#ifndef LIBNITROKEY_NK_STRNDUP_H
+#define LIBNITROKEY_NK_STRNDUP_H
+
+#ifdef __WIN32
+char * strndup(const char* str, size_t maxlen);
+#endif
+
+
+#endif // LIBNITROKEY_NK_STRNDUP_H
